test: Move tests of common utilities into a separate file
These tests were previously incorrectly part of the cli test suite.
